+++ components/marketing/site-footer.tsx
@@ -67,7 +67,7 @@ export function SiteFooter() {
               <Link href="/#product">Product</Link>
               <Link href="/tools">Tools</Link>
               <Link href="/guide">Guide</Link>
-              <Link href="/#pricing">Pricing</Link>
+              <Link href="/pricing">Pricing</Link>
               <Link href="/#faq">FAQ</Link>
             </div>
             <div>